refactor: Split DBHelper into PlaceDBHelper and SearchHistoryDBHelper
- Extracted PlaceDBHelper for managing the PlaceTable database operations.
- Extracted SearchHistoryDBHelper for managing the SearchHistoryTable database operations.
- Deleted DBHelper.kt which previously contained combined functionality.
